+++ astropy/units/format/base.py
@@ -1,4 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+import os
+
 from ...utils.misc import InheritDocstrings
 
 
@@ -16,6 +19,20 @@ class _FormatterMeta(InheritDocstrings):
         mcls.registry[formatter_name] = cls
 
         return cls
+
+
+TAB_HEADER = """# -*- coding: utf-8 -*-
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+# This file was automatically generated from ply. To re-generate this file,
+# remove it from this folder, then build astropy and run the tests in-place:
+#
+#   python setup.py build_ext --inplace
+#   pytest astropy/units
+#
+# You can then commit the changes to this file.
+
+"""
 
 
 class Base(metaclass=_FormatterMeta):
@@ -46,3 +63,15 @@ class Base(metaclass=_FormatterMeta):
 
         raise NotImplementedError(
             "Can not output in {0} format".format(cls.__name__))
+
+    @classmethod
+    def _add_tab_header(cls, name):
+
+        lextab_file = os.path.join(os.path.dirname(__file__), name + '.py')
+
+        with open(lextab_file, 'r') as f:
+            contents = f.read()
+
+        with open(lextab_file, 'w') as f:
+            f.write(TAB_HEADER)
+            f.write(contents)